+++ bmtrain/pipe_layer.py
@@ -5,12 +5,15 @@ import copy
 from typing import Dict, Iterable, Iterator, Tuple, Union, List
 import torch
 
-from .distributed import all_gather, broadcast, all_reduce, send_activations, recv_activations
+from .distributed import all_gather, broadcast, all_reduce, send_activations, recv_activations 
 from .global_var import config
 from . import nccl
-from .checkpointing import ScopedTensorInspectorContext
+from .checkpointing import (
+        ScopedTensorInspectorContext,
+        CheckpointBlockContext
+)
 from . import debug
-from .block_layer import CheckpointBlockContext, CheckpointBlock, round_up, _get_param_kw
+from .block_layer import CheckpointBlock, round_up, _get_param_kw
 
 class OpMicroForward(torch.autograd.Function):
     @staticmethod
@@ -141,6 +144,7 @@ class OpMicroForward(torch.autograd.Function):
                                 [grad_hidden_state],
                             )
                         grad_hidden_state = ipt.grad
+
                         if grad_middle is not None:
                             grad_hidden_state = grad_hidden_state + grad_middle[idx]
                     if ctx.micro_idx == config["micros"]-1:
@@ -310,6 +314,98 @@ class OpPipeTransformerBlockList(torch.autograd.Function):
 
         return (None, None, None, grad) + tuple(grads) + (None,)
 
+def zero_pre_forward(module, inputs):
+    if module._micro_idx == 0:
+        forward_flag = 1 if config['zero_level'] == 2 else 0
+        module._forward_block_ctx = CheckpointBlockContext(module, module._layer_dict, forward_flag, pipe=True)
+        module._forward_block_ctx.enter()
+
+def zero_post_forward(module, inputs, outputs):
+    if module._micro_idx == config['micros'] - 1:
+        module._forward_block_ctx.exit()
+
+
+def pipe_pre_forward(module, inputs):
+    if not module._is_first_stage:
+        if module._is_first_layer:
+            pre_inputs = recv_activations(module.stage_id - 1, config['pipe_comm'])
+            pre_inputs.requires_grad_()
+            return (pre_inputs, ) + inputs[1:]
+
+def pipe_post_forward(module, inputs, outputs):
+    if not module._is_last_stage:
+        if module._is_last_layer:
+            send_data = outputs[0] if isinstance(outputs, tuple) else outputs
+            send_activations(send_data, module.stage_id + 1, config['pipe_comm'])
+
+def zero_pre_backward(module, grad_inputs):
+    if module._micro_idx == config['micros'] - 1:
+        backward_flag = 2 if config['zero_level'] == 2 else 0
+        with torch.enable_grad():
+            module._backward_block_ctxs[module._layer_id] = CheckpointBlockContext(module, module._layer_dict, backward_flag, pipe=True)
+            module._backward_block_ctxs[module._layer_id].enter(True)
+#    if module._micro_idx == 0:
+#        if not module._is_last_layer:
+#            module._backward_block_ctxs[module._layer_id + 1].exit(True)
+
+def zero_post_backward(module, grad_inputs, grad_outputs):
+    if module._micro_idx == 0:
+        module._backward_block_ctxs[module._layer_id].exit(True)
+#        if module._is_first_layer:
+#            module._backward_block_ctxs[module._layer_id].exit(True)
+
+def pipe_pre_backward(module, grad_inputs):
+    if module._is_last_layer:
+        module._grad_list = all_gather(grad_inputs[0], config["pipe_comm"])
+        module._grad_list = module._grad_list.flatten(start_dim=0, end_dim=1).chunk(module.stages, dim=0)
+
+    if module._is_last_layer and module._is_last_stage:
+        return (module._grad_list[module._micro_idx], )
+
+    if not module._is_last_stage:
+        if module._is_last_layer:
+            pre_grad_inputs = recv_activations(module.stage_id + 1, config['pipe_comm'])
+            return (pre_grad_inputs, ) + grad_inputs[1:]
+            
+
+def pipe_post_backward(module, grad_inputs, grad_outputs):
+    if not module._is_first_stage:
+        if module._is_first_layer:
+            send_data = grad_inputs[0] if isinstance(grad_inputs, tuple) else grad_inputs 
+            if send_data is not None:
+                send_activations(send_data, module.stage_id - 1, config['pipe_comm'])
+
+#    if module._is_first_layer:
+#        if module._micro_idx == config['micros'] -1:
+#            module._all_grads = []
+#        grad = grad_inputs[0] 
+#        module._all_grads.append(grad)
+#        if module._micro_idx == 0:
+#            grads = torch.cat(module._all_grads, dim=0)
+#            grad = broadcast(grads, 0, config['pipe_comm'])
+#            grad = grad.chunk(module.stages, dim=0)
+#            return (grad[module.stage_id], ) + grad_inputs[1:]
+
+    module._micro_idx -= 1
+    
+def checkpoint_pre_forward(module, inputs):
+    if module._micro_idx == 0:
+        module._inputs = [inputs]
+        module._cuda_rng_state = [torch.cuda.get_rng_state()]
+    else:
+        module._inputs.append(inputs)
+        module._cuda_rng_state.append(torch.cuda.get_rng_state())
+
+def checkpoint_pre_backward(module, grad_outputs):
+    with torch.random.fork_rng(devices=[torch.cuda.current_device()], enabled=True):
+        with torch.enable_grad():
+            torch.cuda.set_rng_state(module._cuda_rng_state[module._micro_idx])
+            out = module._module(*module._inputs[module._micro_idx])
+            torch.autograd.backward(out, *grad_outputs)
+
+            zero_post_backward(module, None, grad_outputs)
+            pipe_post_backward(module, module._inputs[module._micro_idx][0].grad, None)
+
 class PipelineTransformerBlockList(torch.nn.Module):
     r"""
     TransformerBlockList is a list of CheckpointBlocks.
@@ -331,9 +427,9 @@ class PipelineTransformerBlockList(torch.nn.Module):
     """
     _modules: Dict[str, CheckpointBlock]
 
-    def __init__(self, modules: Iterable[CheckpointBlock]) -> None:
+    def __init__(self, modules: Iterable[CheckpointBlock], num_hidden=1) -> None:
         super().__init__()
-        
+        self.num_hidden = num_hidden 
         self._modules = {}
         rank = config['rank']
         topo = config['topology']
@@ -345,9 +441,39 @@ class PipelineTransformerBlockList(torch.nn.Module):
         for idx, module in enumerate(modules):
             if not isinstance(module, CheckpointBlock):
                 module = CheckpointBlock(module)
+
+            module.register_forward_pre_hook(pipe_pre_forward)
+            module.register_forward_pre_hook(zero_pre_forward)
+            if config['use_checkpoint']:
+                module.register_forward_pre_hook(checkpoint_pre_forward)
+
+            module.register_forward_hook(zero_post_forward)
+            module.register_forward_hook(pipe_post_forward)
+
+            module.register_full_backward_pre_hook(pipe_pre_backward)
+            module.register_full_backward_pre_hook(zero_pre_backward)
+            if config['use_checkpoint']:
+                module.register_full_backward_pre_hook(checkpoint_pre_backward)
+
+            module.register_full_backward_hook(zero_post_backward)
+            module.register_full_backward_hook(pipe_post_backward)
+
+            module.stage_id = self.stage_id
+            module.stages = self.stages
+
             self._modules[str(idx)] = module
 
         self.layer_ids = self.get_range_by_stage_id(self.stage_id)
+
+        self._backward_block_ctxs = [None for _ in range(len(modules))]
+        for i,layer_id in enumerate(self.layer_ids):
+            self._modules[str(layer_id)]._layer_id = layer_id
+            self._modules[str(layer_id)]._backward_block_ctxs = self._backward_block_ctxs 
+            self._modules[str(layer_id)]._is_first_stage = True if self.stage_id == 0 else False
+            self._modules[str(layer_id)]._is_last_stage = True if self.stage_id == self.stages-1 else False
+            self._modules[str(layer_id)]._is_first_layer = True if i == 0 else False
+            self._modules[str(layer_id)]._is_last_layer = True if i == len(self.layer_ids)-1 else False
+
         self.partition_modules(self.layer_ids)
         self.next_rank = pipe_group[self.pipe_idx, self.stage_id + 1] if self.stage_id < config['pipe_size'] - 1 else -1
         self.prev_rank = pipe_group[self.pipe_idx, self.stage_id - 1] if self.stage_id > 0 else -1
@@ -366,15 +492,52 @@ class PipelineTransformerBlockList(torch.nn.Module):
 
     def forward(self, hidden_state, *args, batch_related=[], return_hidden_states=False):
         self.return_hidden_states = return_hidden_states
-        placeholder = torch.tensor([], requires_grad=torch.is_grad_enabled())
-        args = list(args)
-        args.append(batch_related)
-        outputs = OpPipeTransformerBlockList.apply(placeholder, self, self.save_list, hidden_state, *args)
-        hidden_state, middle_states = outputs[:2]
-        if return_hidden_states:
-            return hidden_state, middle_states
+        if False:
+            placeholder = torch.tensor([], requires_grad=torch.is_grad_enabled())
+            args = list(args)
+            args.append(batch_related)
+            outputs = OpPipeTransformerBlockList.apply(placeholder, self, self.save_list, hidden_state, *args)
+            hidden_state, middle_states = outputs[:2]
+            if return_hidden_states:
+                return hidden_state, middle_states
+            else:
+                return hidden_state
         else:
-            return hidden_state
+            batch_size = hidden_state.shape[0]
+            num_micros = config["micros"]
+            hidden_state_list = all_gather(hidden_state, config["pipe_comm"]).flatten(0, 1).detach().requires_grad_()
+
+            args_list = [[] for _ in range(num_micros)]
+            for arg in args:
+                if torch.is_tensor(arg):
+                    arg_all = all_gather(arg, config['pipe_comm'])
+                    if arg.shape[0] == batch_size:
+                        arg_all = arg_all.flatten(0, 1).chunk(num_micros, dim=0)
+                        arg_all = [tensor.detach().requires_grad_(arg.requires_grad) for tensor in arg_all]
+                    else:
+                        arg_all = [arg_all[0].detach().requires_grad_(arg.requires_grad) for i in range(num_micros)]
+                else:
+                    arg_all = [arg for _ in range(num_micros)]
+                for i in range(num_micros):
+                    args_list[i].append(arg_all[i])
+            
+            outputs = []
+            hidden_state_list = hidden_state_list.chunk(num_micros, dim=0)
+            for micro_idx, (hidden_state, arg) in enumerate(zip(hidden_state_list, args_list)):
+                for idx,layer_id in enumerate(self.layer_ids):
+                    self._modules[str(layer_id)]._micro_idx = micro_idx
+                    hidden_state = self._modules[str(layer_id)](hidden_state, *arg)
+                outputs.append(hidden_state)
+
+            last_hidden = torch.cat(outputs, dim=0)
+            last_hidden = broadcast(last_hidden, config["pipe_size"] - 1, config["pipe_comm"])
+            last_hidden = last_hidden.chunk(self.stages, dim=0)
+            outputs = last_hidden[self.stage_id]
+
+        if return_hidden_states:
+            return outputs[:2*self.num_hidden]
+        else:
+            return outputs[:self.num_hidden] if self.num_hidden > 1 else outputs
 
     def get_range_by_stage_id(self, stage_id : int) -> List[int]:
         part_lens = [0]+[self.get_part_len_by_stage_id(i) for i in range(stage_id+1)]
